refactor: 검증을 하기 위한 클래스 생성
- 시도 횟수에 관한 검증을 책임질 AttemptValidator
- 자동차 컬렉션에 관한 검증을책임질 CarsValidator
- 자동차 이름에 관한 검증을 책임질 CarValidator
- 공통된 검증 로직을 담당할 Validator
