+++ src/main/java/racingcar/domain/validation/Validator.java
@@ -1,0 +1,16 @@
+package racingcar.domain.validation;
+
+public class Validator {
+	private static final String NULL_ERROR = "null";
+	private static final String EMPTY_ERROR = "공백";
+	private static final String EMPTY = "";
+
+	public void checkBlank(String name, String text) {
+		if (name == null) {
+			throw new NullPointerException(String.format(text, NULL_ERROR));
+		}
+		if (name.trim().equals(EMPTY)) {
+			throw new IllegalArgumentException(String.format(text, EMPTY_ERROR));
+		}
+	}
+}
